requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
ccxt
airflow-provider-duckdb==0.2.0
astro-sdk-python==1.8.0
duckdb==0.7.1
# minio==7.1.13
pandas==1.5.3
streamlit==1.23.0
confluent-kafka==2.1.1
apache-airflow-providers-apache-kafka==1.0.0